% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/c_1.R
\name{c_1}
\alias{c_1}
\title{Find the location of the smaller maximum of the generic term in the dual.}
\usage{
c_1(eta, gamma, lambda, logscale = 0)
}
\arguments{
\item{eta,gamma}{prior mean and standard error}

\item{lambda}{the dual variable}

\item{logscale}{(optional) is lambda in log-scale? default = 0}
}
\description{
Find the location of the smaller maximum of the generic term in the dual.
}

