% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_crossing.R
\name{find_crossing}
\alias{find_crossing}
\title{find the crossing point of the two competing maxima}
\usage{
find_crossing(eta, gamma)
}
\arguments{
\item{eta}{mean}

\item{gamma}{marginal variance: gamma^2 = 1 + sigma^2, so gamma must be greater than 1}
}
\description{
find the crossing point of the two competing maxima
}

