% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/g_fun.R
\name{g_fun}
\alias{g_fun}
\title{compute the dual constraint function: 1 if lambda <= l_prime, and a phi(c_1) otherwise}
\usage{
g_fun(eta, gamma, lambda, l_prime)
}
\arguments{
\item{eta,gamma}{prior mean and standard error (all inputs must be vectors of the same length)}

\item{lambda}{the dual variable (a scalar)}

\item{l_prime}{crossing points}
}
\value{
Value of the dual constraint (vector of same length as the inputs)
}
\description{
compute the dual constraint function: 1 if lambda <= l_prime, and a phi(c_1) otherwise
}

