% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iGWAS.R
\name{iGWAS}
\alias{iGWAS}
\title{informed Genome-Wide Association Study}
\usage{
iGWAS(P_current, N_current, P_prior, N_prior, q = 0.05, phi = 1)
}
\arguments{
\item{P_current}{P-values in the current study, a numeric vector of length J, with entries between 0 and 1}

\item{N_current}{sample size in the current study, a positive integer (or vector of length J)}

\item{P_prior}{P-values in the prior study, a numeric vector of length J, with entries between 0 and 1}

\item{N_prior}{sample size in the current study, a positive integer (or vector of length J)}

\item{q}{(optional) uncorrected level at which tests should be performed. Default \code{q = 0.05}}

\item{phi}{(optional) dispersion factor used to multiply all standard errors. Default \code{phi = 1}}
}
\value{
A vector of 0-1s indicating the significant tests (1-s)
}
\description{
For more details, see the paper "Optimal Multiple Testing Under a
 Gaussian Prior on the Effect Sizes", by Dobriban, Fortney, Kim and Owen,
  \url{http://arxiv.org/abs/1504.02935}
}

