% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pweight.R
\docType{package}
\name{pweight}
\alias{pweight}
\alias{pweight-package}
\title{pweight: P-value weighting in multiple testing.}
\description{
The pweight package provides functions to compute various p-value
weighting schemes, as Spjotvoll, exponential and Bayes weights. These are methods for improving power in multiple testing via the use of prior informaton.
}
\details{
Several of these methods were proposed by the authors in the following papers:
"Optimal Multiple Testing Under a
 Gaussian Prior on the Effect Sizes", by Dobriban, Fortney, Kim and Owen,
  \url{http://arxiv.org/abs/1504.02935}
}
\section{Weighting functions}{

\code{\link{bayes_weights}} computes Bayes p-value weights

\code{\link{spjotvoll_weights}} computes the Spjotvoll p-value weights

\code{\link{colSums}} computes the exponential weights
}

